:truck: :lock: move analytics frontend to private
