+++ angular.json
@@ -67,32 +67,6 @@
     "elements-layout-page-convl": "libs/elements/layout/page-convl",
     "elements-layout-user-menu": "libs/elements/layout/user-menu",
     "elements-nav-convl-breadcrumbs": "libs/elements/nav/convl/breadcrumbs",
-    "features-analytics": {
-      "projectType": "library",
-      "root": "libs/features/convs-mgr/analytics",
-      "sourceRoot": "libs/features/analytics/src",
-      "prefix": "app",
-      "targets": {
-        "test": {
-          "executor": "@nrwl/jest:jest",
-          "outputs": ["coverage/libs/features/analytics"],
-          "options": {
-            "jestConfig": "libs/features/analytics/jest.config.ts",
-            "passWithNoTests": true
-          }
-        },
-        "lint": {
-          "executor": "@nrwl/linter:eslint",
-          "options": {
-            "lintFilePatterns": [
-              "libs/features/analytics/**/*.ts",
-              "libs/features/analytics/**/*.html"
-            ]
-          }
-        }
-      },
-      "tags": []
-    },
     "features-app-auth-login": "libs/features/app/auth/login",
     "features-convs-mgr-conversations-assessments": {
       "projectType": "library",
@@ -1043,6 +1017,32 @@
           "options": {
             "jestConfig": "libs/model/user/bot-users/jest.config.ts",
             "passWithNoTests": true
+          }
+        }
+      },
+      "tags": []
+    },
+    "private-features-convs-mgr-analytics": {
+      "projectType": "library",
+      "root": "libs/private/features/convs-mgr/analytics",
+      "sourceRoot": "libs/private/features/convs-mgr/analytics/src",
+      "prefix": "app",
+      "targets": {
+        "test": {
+          "executor": "@nrwl/jest:jest",
+          "outputs": ["coverage/libs/private/features/convs-mgr/analytics"],
+          "options": {
+            "jestConfig": "libs/private/features/convs-mgr/analytics/jest.config.ts",
+            "passWithNoTests": true
+          }
+        },
+        "lint": {
+          "executor": "@nrwl/linter:eslint",
+          "options": {
+            "lintFilePatterns": [
+              "libs/private/features/convs-mgr/analytics/**/*.ts",
+              "libs/private/features/convs-mgr/analytics/**/*.html"
+            ]
           }
         }
       },
