--- libs/features/convs-mgr/analytics/README.md
@@ -1,7 +1,0 @@
-# features-analytics
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test features-analytics` to execute the unit tests.
