+++ libs/private/features/convs-mgr/analytics/README.md
@@ -1,0 +1,7 @@
+# private-features-convs-mgr-analytics
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test private-features-convs-mgr-analytics` to execute the unit tests.
