+++ libs/private/features/convs-mgr/analytics/jest.config.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
 export default {
-  displayName: 'features-analytics',
-  preset: '../../../jest.preset.js',
+  displayName: 'private-features-convs-mgr-analytics',
+  preset: '../../../../../jest.preset.js',
   setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
   globals: {
     'ts-jest': {
@@ -9,7 +9,8 @@ export default {
       stringifyContentPathRegex: '\\.(html|svg)$',
     },
   },
-  coverageDirectory: '../../../coverage/libs/features/analytics',
+  coverageDirectory:
+    '../../../../../coverage/libs/private/features/convs-mgr/analytics',
   transform: {
     '^.+\\.(ts|mjs|js|html)$': 'jest-preset-angular',
   },
