+++ libs/private/features/convs-mgr/analytics/src/lib/analytics.module.ts
@@ -8,9 +8,10 @@ import { ReportsPageComponent } from './pages/reports-page/reports-page.componen
 
 import { AnalyticsRouterModule } from './analytics.router.module';
 import { SafePipe } from './pages/safe.pipe';
+import { AnalyticsPageComponent } from './pages/analytics-page/analytics-page.component';
 
 @NgModule({
   imports: [CommonModule, AnalyticsRouterModule, ConvlPageModule],
-  declarations: [DashboardPageComponent, ReportsPageComponent],
+  declarations: [DashboardPageComponent, ReportsPageComponent, AnalyticsPageComponent, SafePipe],
 })
 export class AnalyticsModule {}
