+++ libs/private/state/analytics/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/private-state-analytics.module';
+export * from './lib/services/metabase.service';
