+++ tsconfig.base.json
@@ -259,6 +259,9 @@
       "@app/model/organisation": ["libs/model/organisation/main/src/index.ts"],
       "@app/model/user": ["libs/model/user/base/src/index.ts"],
       "@app/model/user/bot-users": ["libs/model/user/bot-users/src/index.ts"],
+      "@app/private/features/convs-mgr/analytics": [
+        "libs/private/features/convs-mgr/analytics/src/index.ts"
+      ],
       "@app/private/functions/analytics/metabase": [
         "libs/private/functions/analytics/metabase/src/index.ts"
       ],
